authentification terminée

+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UtilisateurController;
+use App\Http\Controllers\AuthController;
 
 /*
 |--------------------------------------------------------------------------
@@ -11,23 +12,25 @@ use App\Http\Controllers\UtilisateurController;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
+
+Route::post('/login', [AuthController::class, 'login']);
+Route::middleware('auth:api')->get('/me', [AuthController::class, 'me']);
+Route::middleware('auth:api')->post('/logout', [AuthController::class, 'logout']);
+
 
 
 
 // Route pour récupérer la liste des utilisateurs
-Route::get('/users', [UtilisateurController::class, 'getUsers'])->name('api.users.index');
+Route::middleware('auth:api')->get('/users', [UtilisateurController::class, 'getUsers'])->name('api.users.index');
 
 // Route pour enregistrer un nouvel utilisateur
-Route::post('users', [UtilisateurController::class, 'enregistrer']);
+Route::middleware('auth:api')->post('users', [UtilisateurController::class, 'enregistrer']);
 
 // Route pour mettre à jour un utilisateur existant
-Route::put('/users/{id}', [UtilisateurController::class, 'update']);
+Route::middleware('auth:api')->put('/users/{id}', [UtilisateurController::class, 'update']);
 
 // Route pour supprimer un utilisateur
-Route::delete('/users/{userId}', [UtilisateurController::class, 'destroy']);
+Route::middleware('auth:api')->delete('/users/{userId}', [UtilisateurController::class, 'destroy']);
 
 // Route pour supprimer plusieurs utilisateurs
-Route::post('/users/suppression-groupe', [UtilisateurController::class, 'suppressionGroupe']);
+Route::middleware('auth:api')->post('/users/suppression-groupe', [UtilisateurController::class, 'suppressionGroupe']);
